Implemented Condition and Notification DAO. Cleaned up unnecessary files

+++ src/main/java/nz/co/twg/eim/model/condition/FileCondition.java
@@ -8,11 +8,11 @@ import java.util.List;
 public class FileCondition implements Condition<List<File>> {
 
     private final String id;
-    private final long maxAge;
+    private final int maxAge;
     private final int fileCount;
     private final String directory;
 
-    public FileCondition(String id, long maxAge, int fileCount, String directory) {
+    public FileCondition(String id, int maxAge, int fileCount, String directory) {
         this.id = id;
         this.maxAge = maxAge;
         this.fileCount = fileCount;
